requirements.txt
langgraph
google-generativeai
python-dotenv
trafilatura
pypdf
requests
duckduckgo-search
playwright
selenium
webdriver-manager
fpdf2
fastapi
uvicorn
websockets
python-multipart
